Restructure tests and initialization to allow tests to work correctly.
